+++ es6/tests/e2e/text.js
@@ -137,4 +137,14 @@ describe("Text templating", () => {
 			wrapMultiError(expectedError)
 		);
 	});
+
+	describe("should work asynchronously", async () => {
+		const doc = new TxtTemplater("Hello {name}", {});
+		const result = await doc.renderAsync({
+			name: new Promise((resolve) => {
+				resolve("John");
+			}),
+		});
+		expect(result).to.equal("Hello John");
+	});
 });